# docs/swagger.yaml
basePath: /
definitions:
  auth.Credentials:
    properties:
      password:
        example: strong_password_123
        type: string
      phone_number:
        example: "+998901234567"
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: REST API для управления заказами, продуктами и пользователями.
  license:
    name: MIT
  termsOfService: http://example.com/terms/
  title: Order Service API
  version: "1.0"
paths:
...
  /api/v1/auth/signin:
    post:
      consumes:
      - application/json
      description: Аутентифицирует пользователя и выдает JWT токен
      parameters:
      - description: Телефон и пароль
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/auth.Credentials'
      produces:
      - application/json
      responses:
        "200":
          description: JWT токен
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Error initialising session token
          schema:
            type: string
      summary: Вход пользователя
      tags:
      - Auth
  /api/v1/auth/signup:
    post:
      consumes:
      - application/json
      description: Регистрирует нового пользователя по номеру телефона и паролю
      parameters:
      - description: Данные пользователя
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/auth.Credentials'
      produces:
      - application/json
      responses:
        "200":
          description: ID нового пользователя
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Invalid request body
          schema:
            type: string
        "409":
          description: A user with this phone number already exists
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Регистрация нового пользователя
      tags:
      - Auth
...
swagger: "2.0"